engine/pyproject.toml
[tool.poetry]
name = "engine"
version = "0.1.0"
description = ""
authors = ["w40141 <w.40141@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
whisper = {git = "https://github.com/openai/whisper.git"}
torch = "^1.13.1"
torchvision = "^0.14.1"
torchaudio = "^0.13.1"
tqdm = "^4.64.1"
transformers = "^4.25.1"
mecab-python3 = "^1.0.6"
unidic-lite = "^1.0.8"
feedparser = "^6.0.10"
algoliasearch = "^2.6.3"
python-dotenv = "^0.21.0"
ffmpeg-python = "^0.2.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
